backend/requirements.txt

# Django Framework
Django>=5.0.0
djangorestframework>=3.14.0

# JWT Authentication
PyJWT>=2.8.0

# Database connectors
pymongo>=4.5.0
dnspython>=2.4.0

# Environment variables
python-dotenv>=1.0.0

# CORS handling
django-cors-headers>=4.3.0
flask-cors>=4.0.0  

# Password hashing
bcrypt>=4.1.0

# Utilities
pytz>=2023.3
python-dateutil>=2.8.0

# Flask Framework (for Flask implementation)
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Bcrypt>=1.0.0
Flask-Login>=0.6.0
Flask-PyMongo>=2.3.0  
Flask-JWT-Extended>=4.5.0  
Flask-Limiter>=3.5.0  

# AI/ML Dependencies for Smart Planner
sentence-transformers>=2.2.0  # BERT multilingual embeddings
scipy>=1.11.0  # SVD matrix factorization
numpy>=1.24.0  # Vector operations
scikit-learn>=1.3.0  # ML utilities

amadeus
airportsdata